51. N-Queens
